refactor(metal): move the RMSNorm kernels out of elem.rs into norm.rs, no behaviour change
elem.rs 1446 -> 718 lines. The three norm kernel sources, their nine
encoders, the per-head parity test and the norm half of the prefill
pipeline warm-up move together; elem keeps the residual add, the GLU
pairs, axpy and argmax.
